+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/ClientSideStatementValueConverters.java
@@ -16,6 +16,9 @@
 
 package com.google.cloud.spanner.connection;
 
+import static com.google.cloud.spanner.connection.ReadOnlyStalenessUtil.parseTimeUnit;
+import static com.google.cloud.spanner.connection.ReadOnlyStalenessUtil.toChronoUnit;
+
 import com.google.cloud.spanner.ErrorCode;
 import com.google.cloud.spanner.Options.RpcPriority;
 import com.google.cloud.spanner.SpannerException;
@@ -27,16 +30,14 @@ import com.google.cloud.spanner.connection.PgTransactionMode.IsolationLevel;
 import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
-import com.google.protobuf.Duration;
-import com.google.protobuf.util.Durations;
 import com.google.spanner.v1.DirectedReadOptions;
-import com.google.spanner.v1.RequestOptions.Priority;
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
 import java.util.Base64;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -166,9 +167,16 @@ class ClientSideStatementValueConverters {
 
   /** Converter from string to {@link Duration}. */
   static class DurationConverter implements ClientSideStatementValueConverter<Duration> {
+    private final String resetValue;
+
     private final Pattern allowedValues;
 
     public DurationConverter(String allowedValues) {
+      this("NULL", allowedValues);
+    }
+
+    DurationConverter(String resetValue, String allowedValues) {
+      this.resetValue = Preconditions.checkNotNull(resetValue);
       // Remove the parentheses from the beginning and end.
       this.allowedValues =
           Pattern.compile(
@@ -184,17 +192,25 @@ class ClientSideStatementValueConverters {
     public Duration convert(String value) {
       Matcher matcher = allowedValues.matcher(value);
       if (matcher.find()) {
-        if (matcher.group(0).equalsIgnoreCase("null")) {
-          return Durations.fromNanos(0L);
+        if (value.trim().equalsIgnoreCase(resetValue)) {
+          return Duration.ZERO;
         } else {
-          Duration duration =
-              ReadOnlyStalenessUtil.createDuration(
-                  Long.parseLong(matcher.group(1)),
-                  ReadOnlyStalenessUtil.parseTimeUnit(matcher.group(2)));
-          if (duration.getSeconds() == 0L && duration.getNanos() == 0) {
+          try {
+            Duration duration;
+            if (matcher.group(1) != null && matcher.group(2) != null) {
+              ChronoUnit unit = toChronoUnit(parseTimeUnit(matcher.group(2)));
+              duration = Duration.of(Long.parseLong(matcher.group(1)), unit);
+            } else {
+              duration = Duration.ofMillis(Long.parseLong(value.trim()));
+            }
+            if (duration.isZero()) {
+              return null;
+            }
+            return duration;
+          } catch (NumberFormatException exception) {
+            // Converters should return null for invalid values.
             return null;
           }
-          return duration;
         }
       }
       return null;
@@ -202,44 +218,9 @@ class ClientSideStatementValueConverters {
   }
 
   /** Converter from string to {@link Duration}. */
-  static class PgDurationConverter implements ClientSideStatementValueConverter<Duration> {
-    private final Pattern allowedValues;
-
+  static class PgDurationConverter extends DurationConverter {
     public PgDurationConverter(String allowedValues) {
-      // Remove the parentheses from the beginning and end.
-      this.allowedValues =
-          Pattern.compile(
-              "(?is)\\A" + allowedValues.substring(1, allowedValues.length() - 1) + "\\z");
-    }
-
-    @Override
-    public Class<Duration> getParameterClass() {
-      return Duration.class;
-    }
-
-    @Override
-    public Duration convert(String value) {
-      Matcher matcher = allowedValues.matcher(value);
-      if (matcher.find()) {
-        Duration duration;
-        if (matcher.group(0).equalsIgnoreCase("default")) {
-          return Durations.fromNanos(0L);
-        } else if (matcher.group(2) == null) {
-          duration =
-              ReadOnlyStalenessUtil.createDuration(
-                  Long.parseLong(matcher.group(0)), TimeUnit.MILLISECONDS);
-        } else {
-          duration =
-              ReadOnlyStalenessUtil.createDuration(
-                  Long.parseLong(matcher.group(1)),
-                  ReadOnlyStalenessUtil.parseTimeUnit(matcher.group(2)));
-        }
-        if (duration.getSeconds() == 0L && duration.getNanos() == 0) {
-          return null;
-        }
-        return duration;
-      }
-      return null;
+      super("DEFAULT", allowedValues);
     }
   }
 
@@ -289,7 +270,7 @@ class ClientSideStatementValueConverters {
             try {
               return TimestampBound.ofExactStaleness(
                   Long.parseLong(matcher.group(groupIndex + 2)),
-                  ReadOnlyStalenessUtil.parseTimeUnit(matcher.group(groupIndex + 3)));
+                  parseTimeUnit(matcher.group(groupIndex + 3)));
             } catch (IllegalArgumentException e) {
               throw SpannerExceptionFactory.newSpannerException(
                   ErrorCode.INVALID_ARGUMENT, e.getMessage());
@@ -298,7 +279,7 @@ class ClientSideStatementValueConverters {
             try {
               return TimestampBound.ofMaxStaleness(
                   Long.parseLong(matcher.group(groupIndex + 2)),
-                  ReadOnlyStalenessUtil.parseTimeUnit(matcher.group(groupIndex + 3)));
+                  parseTimeUnit(matcher.group(groupIndex + 3)));
             } catch (IllegalArgumentException e) {
               throw SpannerExceptionFactory.newSpannerException(
                   ErrorCode.INVALID_ARGUMENT, e.getMessage());
@@ -499,9 +480,9 @@ class ClientSideStatementValueConverters {
   }
 
   /** Converter for converting strings to {@link RpcPriority} values. */
-  static class RpcPriorityConverter implements ClientSideStatementValueConverter<Priority> {
-    private final CaseInsensitiveEnumMap<Priority> values =
-        new CaseInsensitiveEnumMap<>(Priority.class);
+  static class RpcPriorityConverter implements ClientSideStatementValueConverter<RpcPriority> {
+    private final CaseInsensitiveEnumMap<RpcPriority> values =
+        new CaseInsensitiveEnumMap<>(RpcPriority.class);
     private final Pattern allowedValues;
 
     public RpcPriorityConverter(String allowedValues) {
@@ -512,19 +493,19 @@ class ClientSideStatementValueConverters {
     }
 
     @Override
-    public Class<Priority> getParameterClass() {
-      return Priority.class;
+    public Class<RpcPriority> getParameterClass() {
+      return RpcPriority.class;
     }
 
     @Override
-    public Priority convert(String value) {
+    public RpcPriority convert(String value) {
       Matcher matcher = allowedValues.matcher(value);
       if (matcher.find()) {
         if (matcher.group(0).equalsIgnoreCase("null")) {
-          return Priority.PRIORITY_UNSPECIFIED;
+          return RpcPriority.UNSPECIFIED;
         }
       }
-      return values.get("PRIORITY_" + value);
+      return values.get(value);
     }
   }