+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/spi/v1/GapicSpannerRpc.java
@@ -393,6 +393,8 @@ public class GapicSpannerRpc implements SpannerRpc {
               .withCheckInterval(checkInterval)
               .withClock(NanoClock.getDefaultClock());
 
+      final String emulatorHost = System.getenv("SPANNER_EMULATOR_HOST");
+
       try {
         this.spannerStub =
             GrpcSpannerStub.create(
@@ -402,7 +404,9 @@ public class GapicSpannerRpc implements SpannerRpc {
                     .setTransportChannelProvider(channelProvider)
                     .setCredentialsProvider(credentialsProvider)
                     .setStreamWatchdogProvider(watchdogProvider)
-                    .setTracerFactory(options.getApiTracerFactory())
+                    .setTracerFactory(
+                        options.getApiTracerFactory(
+                            /* isAdminClient = */ false, isEmulatorEnabled(options, emulatorHost)))
                     .build());
         this.readRetrySettings =
             options.getSpannerStubSettings().streamingReadSettings().getRetrySettings();
@@ -430,7 +434,9 @@ public class GapicSpannerRpc implements SpannerRpc {
             .setTransportChannelProvider(channelProvider)
             .setCredentialsProvider(credentialsProvider)
             .setStreamWatchdogProvider(watchdogProvider)
-            .setTracerFactory(options.getApiTracerFactory())
+            .setTracerFactory(
+                options.getApiTracerFactory(
+                    /* isAdminClient = */ false, isEmulatorEnabled(options, emulatorHost)))
             .executeSqlSettings()
             .setRetrySettings(partitionedDmlRetrySettings);
         pdmlSettings.executeStreamingSqlSettings().setRetrySettings(partitionedDmlRetrySettings);
@@ -457,7 +463,9 @@ public class GapicSpannerRpc implements SpannerRpc {
                 .setTransportChannelProvider(channelProvider)
                 .setCredentialsProvider(credentialsProvider)
                 .setStreamWatchdogProvider(watchdogProvider)
-                .setTracerFactory(options.getApiTracerFactory())
+                .setTracerFactory(
+                    options.getApiTracerFactory(
+                        /* isAdminClient = */ true, isEmulatorEnabled(options, emulatorHost)))
                 .build();
         this.instanceAdminStub = GrpcInstanceAdminStub.create(instanceAdminStubSettings);
 
@@ -468,7 +476,9 @@ public class GapicSpannerRpc implements SpannerRpc {
                 .setTransportChannelProvider(channelProvider)
                 .setCredentialsProvider(credentialsProvider)
                 .setStreamWatchdogProvider(watchdogProvider)
-                .setTracerFactory(options.getApiTracerFactory())
+                .setTracerFactory(
+                    options.getApiTracerFactory(
+                        /* isAdminClient = */ true, isEmulatorEnabled(options, emulatorHost)))
                 .build();
 
         // Automatically retry RESOURCE_EXHAUSTED for GetOperation if auto-throttling of
@@ -512,7 +522,7 @@ public class GapicSpannerRpc implements SpannerRpc {
 
         // Check whether the SPANNER_EMULATOR_HOST env var has been set, and if so, if the emulator
         // is actually running.
-        checkEmulatorConnection(options, channelProvider, credentialsProvider);
+        checkEmulatorConnection(options, channelProvider, credentialsProvider, emulatorHost);
       } catch (Exception e) {
         throw newSpannerException(e);
       }
@@ -611,15 +621,11 @@ public class GapicSpannerRpc implements SpannerRpc {
   private static void checkEmulatorConnection(
       SpannerOptions options,
       TransportChannelProvider channelProvider,
-      CredentialsProvider credentialsProvider)
+      CredentialsProvider credentialsProvider,
+      String emulatorHost)
       throws IOException {
-    final String emulatorHost = System.getenv("SPANNER_EMULATOR_HOST");
     // Only do the check if the emulator environment variable has been set to localhost.
-    if (options.getChannelProvider() == null
-        && emulatorHost != null
-        && options.getHost() != null
-        && options.getHost().startsWith("http://localhost")
-        && options.getHost().endsWith(emulatorHost)) {
+    if (isEmulatorEnabled(options, emulatorHost)) {
       // Do a quick check to see if the emulator is actually running.
       try {
         InstanceAdminStubSettings.Builder testEmulatorSettings =
@@ -650,6 +656,15 @@ public class GapicSpannerRpc implements SpannerRpc {
                 emulatorHost));
       }
     }
+  }
+
+  private static boolean isEmulatorEnabled(SpannerOptions options, String emulatorHost) {
+    // Only do the check if the emulator environment variable has been set to localhost.
+    return options.getChannelProvider() == null
+        && emulatorHost != null
+        && options.getHost() != null
+        && options.getHost().startsWith("http://localhost")
+        && options.getHost().endsWith(emulatorHost);
   }
 
   private static final RetrySettings ADMIN_REQUESTS_LIMIT_EXCEEDED_RETRY_SETTINGS =